deps/auth/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
faker==15.0.0
    # via -r requirements.in
grpcio==1.49.1
    # via
    #   -r requirements.in
    #   grpcio-tools
grpcio-tools==1.49.1
    # via -r requirements.in
protobuf==4.21.7
    # via grpcio-tools
pydantic==1.10.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via faker
six==1.16.0
    # via
    #   grpcio
    #   python-dateutil
typing-extensions==4.3.0
    # via pydantic

# The following packages are considered to be unsafe in a requirements file:
# setuptools
